+++ app/Http/Requests/ContratodespesaacessoriaRequest.php
@@ -1,0 +1,62 @@
+<?php
+
+namespace App\Http\Requests;
+
+use App\Http\Requests\Request;
+use Illuminate\Foundation\Http\FormRequest;
+
+class ContratodespesaacessoriaRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        // only allow updates if the user is logged in
+        return backpack_auth()->check();
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'tipo_id' => "required",
+            'contrato_id' => "required",
+            'recorrencia_id' => "required",
+            'valor' => "required|min:0.01",
+        ];
+    }
+
+    /**
+     * Get the validation attributes that apply to the request.
+     *
+     * @return array
+     */
+    public function attributes()
+    {
+        return [
+            'tipo_id' => "Tipo Despesa",
+            'contrato_id' => "Contrato",
+            'recorrencia_id' => "Recorrência Despesa",
+            'valor' => "Valor",
+        ];
+    }
+
+    /**
+     * Get the validation messages that apply to the request.
+     *
+     * @return array
+     */
+    public function messages()
+    {
+        return [
+            //
+        ];
+    }
+}